requirements.txt
aiogram==3.4.1
pytesseract
pillow
ultralytics
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
itsdangerous==2.1.2
authlib==1.2.1
httpx==0.25.1
python-dotenv==1.0.0
bcrypt==4.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0.post1
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1